Not a fan of being this tired this early
What the fuck man it was dope to be awake super early but now I'm too fucking tired to do anything at 8 PM? Yeah fuck that man this is bullshit. Anyway I fucked with the damage scripts again and maybe we can talk about that tomorrow idk? Time 4 bed goodnight github. Boy it'd be scary if Github said goodnight back.