8 Directional Dashing Works?
It works now and I fixed some other bugs. BUT I'm not happy with the way diagonal dashing works... It seems like it goes horizontal THEN vertical, rather than both at the same time. Meaning if you diagonally dashed to get out of the way of a bullet you'd slam right the fuck into it...