Grapple crosshair is now PERFECT!
If you point it at something grappleable, it'll zoom out to wherever the grapple with shoot to and turn green
If you point it at something you can't grapple to, it'll turn red and push closer to you if the object is closer than the crosshair.
If you point it at nothing in range, it just stays near the player like before and turns blue.